обернул везде примитивы в обертки

+++ commons/src/main/java/org/parsingbot/repository/UserRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 /**
  * Репозиторий для работы с сущностью User
  */
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     /**
      * @param userId id пользователя
@@ -24,7 +24,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
             "join users on users_vacancies.user_id = users.id " +
             "where users.id = ?1",
             nativeQuery = true)
-    List<Integer> findUserVacanciesIds(long userId);
+    List<Integer> findUserVacanciesIds(Long userId);
 
     /**
      * @param userName имя пользователя
@@ -36,7 +36,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
      * @param chatId id чата с пользователем
      * @return Optional обертка над объектом пользователя
      */
-    Optional<User> findUserByChatId(long chatId);
+    Optional<User> findUserByChatId(Long chatId);
 
     /**
      * @return список пользователей, у которых is_subscribed = true
@@ -51,26 +51,26 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     @Transactional
     @Modifying(clearAutomatically = true)
     @Query(value = "update users set state = ?2 where id = ?1", nativeQuery = true)
-    void updateStateByUserId(long userId, String state);
+    void updateStateByUserId(Long userId, String state);
 
     /**
      * @param userId id пользователя
      * @return имя вакансии для поиска
      */
     @Query(value = "select vacancy_name from users where users.id = ?1", nativeQuery = true)
-    String findVacancyNameByUserId(long userId);
+    String findVacancyNameByUserId(Long userId);
 
     /**
      * @param userId id пользователя
      * @return количество вакансий для поиска
      */
     @Query(value = "select number_of_vacancies from users where users.id = ?1", nativeQuery = true)
-    long findNumberOfVacanciesByUserId(long userId);
+    Long findNumberOfVacanciesByUserId(Long userId);
 
     /**
      * @param userId id пользователя
      * @return ключевые слова в виде строки для поиска вакансий
      */
     @Query(value = "select keywords from users where users.id = ?1", nativeQuery = true)
-    String findKeywordsByUserId(long userId);
+    String findKeywordsByUserId(Long userId);
 }
